add new routing handling to Navigation
