+++ src/pages/floor/[index].js
@@ -1,63 +1,49 @@
-import MoneyDisplay from "~/src/components/MoneyDisplay";
 import Product from "~/src/components/Product";
 import styled from "styled-components";
-import Navigation from "~/src/components/Navigation";
 import Layout from "~/src/components/Layout";
-import { useRouter } from "next/router";
 import useStore from "~/src/zustand/store";
+import BossFloor from "../../components/BossFloor";
 
 export default function HomePage() {
-  const router = useRouter();
-  const { index } = router.query;
   const currentFloor = useStore((state) => state.currentFloor);
 
   function getPosition() {
     switch (currentFloor) {
       case 1:
-        return { floor1: "80px", floor2: "-920px" };
+        return { floor1: "80px", floor2: "-920px", bossFloor: "-920px" };
       case 2:
-        return { floor1: "1080px", floor2: "80px" };
+        return { floor1: "1080px", floor2: "80px", bossFloor: "-920px" };
+      default:
+        return { floor1: "1080px", floor2: "1080px", bossFloor: "80px" };
     }
   }
 
-  const { floor1, floor2 } = getPosition();
+  const { floor1, floor2, bossFloor } = getPosition();
 
-  if (index === "1") {
-    return (
-      <>
-        <Layout>
-          <StyledMain>
-            <StyledSection floor={floor2}>
-              <Product key="tier6" tierId="tier6" />
-              <Product key="tier7" tierId="tier7" />
-              <Product key="tier8" tierId="tier8" />
-              <Product key="tier9" tierId="tier9" />
-            </StyledSection>
-            <StyledSection floor={floor1}>
-              <Product key="tier1" tierId="tier1" />
-              <Product key="tier2" tierId="tier2" />
-              <Product key="tier3" tierId="tier3" />
-              <Product key="tier4" tierId="tier4" />
-              <Product key="tier5" tierId="tier5" />
-            </StyledSection>
-          </StyledMain>
-        </Layout>
-      </>
-    );
-  } else if (index === "2") {
-    return (
-      <>
-        <Layout>
-          <StyledMain>
+  return (
+    <>
+      <Layout>
+        <StyledMain>
+          <StyledSection floor={bossFloor}>
+            <BossFloor />
+          </StyledSection>
+          <StyledSection floor={floor2}>
             <Product key="tier6" tierId="tier6" />
             <Product key="tier7" tierId="tier7" />
             <Product key="tier8" tierId="tier8" />
             <Product key="tier9" tierId="tier9" />
-          </StyledMain>
-        </Layout>
-      </>
-    );
-  }
+          </StyledSection>
+          <StyledSection floor={floor1}>
+            <Product key="tier1" tierId="tier1" />
+            <Product key="tier2" tierId="tier2" />
+            <Product key="tier3" tierId="tier3" />
+            <Product key="tier4" tierId="tier4" />
+            <Product key="tier5" tierId="tier5" />
+          </StyledSection>
+        </StyledMain>
+      </Layout>
+    </>
+  );
 }
 
 const StyledSection = styled.section`
